requirements.txt
# Core Dependencies
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM APIs
google-generativeai==0.3.2
groq==0.4.1

# Retrieval
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-json-logger==2.0.7
python-dotenv==1.0.0
pathlib==1.0.1

# Testing (Required for CI/Validation)
pytest==7.4.4
pytest-asyncio==0.23.3
